refactor: LineInfo -> Line 통합 및 LineInfo 제거
